+++ pkg/platform/microservice/purchaseorderapi/repo.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/dolittle-entropy/platform-api/pkg/dolittle/k8s"
 	"github.com/dolittle-entropy/platform-api/pkg/platform"
-	microserviceK8s "github.com/dolittle-entropy/platform-api/pkg/platform/microservice/k8s"
 	"github.com/dolittle-entropy/platform-api/pkg/platform/microservice/rawdatalog"
 )
 
@@ -23,7 +22,7 @@ func NewRepo(k8sResource K8sResource, rawDataLogIngestorRepo rawdatalog.RawDataL
 }
 
 // Create creates a new PurchaseOrderAPI microservice, and a RawDataLog and WebhookListener if they don't exist.
-func (r *repo) Create(namespace string, tenant k8s.Tenant, application k8s.Application, input platform.HttpInputPurchaseOrderInfo) error {
+func (r *repo) Create(namespace string, customer k8s.Tenant, application k8s.Application, tenant platform.TenantId, input platform.HttpInputPurchaseOrderInfo) error {
 	// TODO not sure where this comes from really, assume dynamic
 
 	environment := input.Environment
@@ -35,10 +34,10 @@ func (r *repo) Create(namespace string, tenant k8s.Tenant, application k8s.Appli
 	microservice := k8s.Microservice{
 		ID:          microserviceID,
 		Name:        microserviceName,
-		Tenant:      tenant,
+		Tenant:      customer,
 		Application: application,
 		Environment: environment,
-		ResourceID:  microserviceK8s.TodoCustomersTenantID,
+		ResourceID:  string(tenant),
 		Kind:        platform.MicroserviceKindPurchaseOrderAPI,
 	}
 
@@ -48,7 +47,7 @@ func (r *repo) Create(namespace string, tenant k8s.Tenant, application k8s.Appli
 	// 	return err
 	// }
 
-	if err := r.k8sResource.Create(namespace, headImage, runtimeImage, microservice, input.Extra, ctx); err != nil {
+	if err := r.k8sResource.Create(namespace, headImage, runtimeImage, microservice, tenant, input.Extra, ctx); err != nil {
 		return err
 	}