+++ pkg/platform/microservice/rawdatalog/repo.go
@@ -6,6 +6,8 @@ import (
 
 	"github.com/dolittle-entropy/platform-api/pkg/dolittle/k8s"
 	"github.com/dolittle-entropy/platform-api/pkg/platform"
+	"github.com/dolittle-entropy/platform-api/pkg/platform/storage"
+	v1 "k8s.io/api/apps/v1"
 	"k8s.io/client-go/kubernetes"
 
 	"context"
@@ -26,12 +28,14 @@ import (
 type RawDataLogIngestorRepo struct {
 	k8sClient       kubernetes.Interface
 	k8sDolittleRepo platform.K8sRepo
+	gitRepo         storage.Repo
 }
 
-func NewRawDataLogIngestorRepo(k8sDolittleRepo platform.K8sRepo, k8sClient kubernetes.Interface) RawDataLogIngestorRepo {
+func NewRawDataLogIngestorRepo(k8sDolittleRepo platform.K8sRepo, k8sClient *kubernetes.Clientset, gitRepo storage.Repo) RawDataLogIngestorRepo {
 	return RawDataLogIngestorRepo{
 		k8sClient:       k8sClient,
 		k8sDolittleRepo: k8sDolittleRepo,
+		gitRepo:         gitRepo,
 	}
 }
 
@@ -44,16 +48,24 @@ func (r RawDataLogIngestorRepo) Update(namespace string, tenant k8s.Tenant, appl
 	return errors.New("TODO")
 }
 
-func (r RawDataLogIngestorRepo) Create(namespace string, tenant k8s.Tenant, application k8s.Application, applicationIngress k8s.Ingress, input platform.HttpInputRawDataLogIngestorInfo) error {
+func (r RawDataLogIngestorRepo) Create(namespace string, customer k8s.Tenant, application k8s.Application, applicationIngress k8s.Ingress, input platform.HttpInputRawDataLogIngestorInfo) error {
+	config := r.k8sDolittleRepo.GetRestConfig()
+	ctx := context.TODO()
+
+	templates := []string{
+		k8sRawDataLogIngestorNats,
+		k8sRawDataLogIngestorStanInMemory,
+	}
 
 	labels := map[string]string{
-		"tenant":      tenant.Name,
-		"application": application.Name,
-		"environment": input.Environment,
+		"tenant":       customer.Name,
+		"application":  application.Name,
+		"environment":  input.Environment,
+		"microservice": input.Name,
 	}
 
 	annotations := map[string]string{
-		"dolittle.io/tenant-id":       tenant.ID,
+		"dolittle.io/tenant-id":       customer.ID,
 		"dolittle.io/application-id":  application.ID,
 		"dolittle.io/microservice-id": input.Dolittle.MicroserviceID,
 	}
@@ -70,7 +82,7 @@ func (r RawDataLogIngestorRepo) Create(namespace string, tenant k8s.Tenant, appl
 	// @joel shoulnd we skip this part and only do the doNats() call?
 
 	// TODO add microservice
-	err := r.doDolittle(namespace, tenant, application, applicationIngress, input)
+	err := r.doDolittle(namespace, customer, application, applicationIngress, input)
 	if err != nil {
 		fmt.Println("Could not doDolittle", err)
 		return err
@@ -304,10 +316,18 @@ func (r RawDataLogIngestorRepo) doNats(namespace string, labels, annotations k8s
 	return nil
 }
 
-func (r RawDataLogIngestorRepo) doDolittle(namespace string, tenant k8s.Tenant, application k8s.Application, applicationIngress k8s.Ingress, input platform.HttpInputRawDataLogIngestorInfo) error {
+func (r RawDataLogIngestorRepo) doDolittle(namespace string, customer k8s.Tenant, application k8s.Application, applicationIngress k8s.Ingress, input platform.HttpInputRawDataLogIngestorInfo) error {
 
 	// TODO not sure where this comes from really, assume dynamic
-	customersTenantID := "17426336-fb8e-4425-8ab7-07d488367be9"
+	// tenantID := "17426336-fb8e-4425-8ab7-07d488367be9"
+	storedApplication, err := r.gitRepo.GetApplication(customer.ID, application.ID)
+	if err != nil {
+		return err
+	}
+	tenantID, err := storedApplication.GetTenantForEnvironment(input.Environment)
+	if err != nil {
+		return err
+	}
 
 	environment := strings.ToLower(input.Environment)
 	host := applicationIngress.Host
@@ -321,10 +341,10 @@ func (r RawDataLogIngestorRepo) doDolittle(namespace string, tenant k8s.Tenant, 
 	microservice := k8s.Microservice{
 		ID:          microserviceID,
 		Name:        microserviceName,
-		Tenant:      tenant,
+		Tenant:      customer,
 		Application: application,
 		Environment: environment,
-		ResourceID:  customersTenantID,
+		ResourceID:  string(tenantID),
 	}
 
 	ingressServiceName := strings.ToLower(fmt.Sprintf("%s-%s", microservice.Environment, microservice.Name))
@@ -343,7 +363,7 @@ func (r RawDataLogIngestorRepo) doDolittle(namespace string, tenant k8s.Tenant, 
 
 	// TODO do I need this?
 	// TODO if I remove it, do I remove the config mapping?
-	microserviceConfigmap := k8s.NewMicroserviceConfigmap(microservice, customersTenantID)
+	microserviceConfigmap := k8s.NewMicroserviceConfigmap(microservice, string(tenantID))
 	deployment := k8s.NewDeployment(microservice, headImage, runtimeImage)
 	service := k8s.NewService(microservice)
 	ingress := k8s.NewIngress(microservice)
@@ -374,7 +394,7 @@ func (r RawDataLogIngestorRepo) doDolittle(namespace string, tenant k8s.Tenant, 
 		"WEBHOOK_REPO":            input.Extra.WriteTo,
 		"LISTEN_ON":               "0.0.0.0:8080",
 		"WEBHOOK_PREFIX":          webhookPrefix,
-		"DOLITTLE_TENANT_ID":      tenant.ID,
+		"DOLITTLE_TENANT_ID":      customer.ID,
 		"DOLITTLE_APPLICATION_ID": application.ID,
 		"DOLITTLE_ENVIRONMENT":    environment,
 		"MICROSERVICE_CONFIG":     "/app/data/microservice_data_from_studio.json",
@@ -401,7 +421,6 @@ func (r RawDataLogIngestorRepo) doDolittle(namespace string, tenant k8s.Tenant, 
 	deployment.Spec.Template.Spec.Containers[0].Ports[0].ContainerPort = 8080
 
 	// Assuming the namespace exists
-	var err error
 	client := r.k8sClient
 	ctx := context.TODO()